* Casting issue resolved in mileage shortage.

+++ ArmsServices/DataServices/Finance/Transactions/MileageShortageReceiptService.cs
@@ -174,7 +174,7 @@ namespace DAL.DataServices.Finance.Transactions
                 MileageShortageReceiptID = dr.GetInt32("MileageShortageReceiptID"),
                 NatureOfTransaction = dr.GetString("NatureOfTransaction"),
                 ReceiptMode = dr.GetString("ReceiptMode"),
-                TripID = dr.GetInt32("TripID"),
+                TripID = dr.GetInt64("TripID"),
                 TripNo = dr.GetString("TripNo"),
                 AssetTransferID = dr.GetInt32("AssetTransferID"),
                 RequestApprovalHistoryID = dr.GetInt32("RequestApprovalHistoryID"),
